Smaller LongObject

+++ Include/cpython/longintrepr.h
@@ -80,7 +80,8 @@ typedef long stwodigits; /* signed variant of twodigits */
 */
 
 typedef struct _PyLongValue {
-    uintptr_t lv_tag; /* Number of digits, sign and flags */
+    // uintptr_t lv_tag; /* Number of digits, sign and flags */
+    uint32_t lv_tag; /* Number of digits, sign and flags */
     digit ob_digit[1];
 } _PyLongValue;
 
@@ -120,7 +121,7 @@ _PyLong_CompactValue(const PyLongObject *op)
 {
     assert(PyType_HasFeature((op)->ob_base.ob_type, Py_TPFLAGS_LONG_SUBCLASS));
     assert(PyUnstable_Long_IsCompact(op));
-    Py_ssize_t sign = 1 - (op->long_value.lv_tag & _PyLong_SIGN_MASK);
+    Py_ssize_t sign = 1ul - (op->long_value.lv_tag & _PyLong_SIGN_MASK);
     return sign * (Py_ssize_t)op->long_value.ob_digit[0];
 }
